requirements.txt
setuptools == 5.5.1.egg
scikit_learn == 0.16.1
PyYAML == 3.11
